use utils in all files
use utility methods in both:

- create_account.php
- login.php
